Factor out `shader.h` from `shader_pass.h`
